requirements.txt
python-telegram-bot>=22.5
APScheduler==3.10.4
python-dotenv==1.0.0
markdown==3.5.2
